Rename vainilla_bs module to vanilla_bs
